* Sort order selectable in settings for dictionary searches and other sorted word listings. * Fixed: Multi line model crash when layout was changed, because updated persistent indexes clashed. * Some changes in the project file might make it not compilable by vs2013. (Although I have been using c++11 features not available there anyway.)

+++ zlineedit.cpp
@@ -74,7 +74,7 @@ bool ZLineEdit::hasAcceptableInput() const
     return validInput();
 }
 
-void ZLineEdit::setCharacterSize(int chsiz)
+void ZLineEdit::setCharacterSize(double chsiz)
 {
     QFontMetrics fm(font());
     int w = fm.averageCharWidth();
